+++ src/dataStructures/graph/minCostConnectPoints.ts
@@ -1,4 +1,8 @@
-import { kruskal, prim } from './utils';
+import { prim } from './utils';
+
+function getWeight(point1: [number, number], point2: [number, number]): number {
+  return Math.abs(point1[0] - point2[0]) + Math.abs(point1[1] - point2[1]);
+}
 
 function buildGraph(points: Array<[number, number]>): number[][] {
   const n = points.length;
@@ -18,26 +22,12 @@ function buildGraph(points: Array<[number, number]>): number[][] {
   return graph;
 }
 
-function getWeight(point1: [number, number], point2: [number, number]): number {
-  return Math.abs(point1[0] - point2[0]) + Math.abs(point1[1] - point2[1]);
-}
-
 // https://leetcode-cn.com/problems/min-cost-to-connect-all-points/
 // 1584
-export function minCostConnectPointsPrim(points: Array<[number, number]>): number {
+export function minCostConnectPoints(points: Array<[number, number]>): number {
   if (points.length <= 1) {
     return 0;
   } else {
     return prim(buildGraph(points));
-  }
-}
-
-// https://leetcode-cn.com/problems/min-cost-to-connect-all-points/
-// 1584
-export function minCostConnectPointsKruskal(points: Array<[number, number]>): number {
-  if (points.length <= 1) {
-    return 0;
-  } else {
-    return kruskal(buildGraph(points));
   }
 }